cleanup, rename appeals folder
